# drakkar/uiserver/openapi.yaml
...
openapi: 3.1.0
info:
  title: Drakkar Backend API
  version: "1.2"
  description: >-
    The JSON surface every Drakkar backend (Python reference and Go port)
    serves identically under /api/v1, plus the unprefixed Kubernetes probes.
    One optional bearer token protects every /api/* route: configure
    ui.auth_token and send `Authorization: Bearer <t>` or `?token=<t>`;
    with no token configured, auth is disabled (private-contour default).


    Not modeled by OpenAPI but part of the contract: `GET /ws` upgrades to a
    WebSocket streaming recorder events — accept, then token check (close
    4401) and origin check (close 4403); at most 100 events per drain pass,
    20 ms idle backoff, one JSON recorder-event object per text frame.


    Legacy unprefixed `/api/...` aliases exist for backends' built-in pages
    and are deliberately outside this document; the SPA only calls /api/v1.
servers:
  - url: /
    description: Same-origin — the backend serves both the SPA and the API.
security:
  - bearerAuth: []
...
paths:
...
  /api/v1/debug/cache/entries:
    get:
      tags: [cache]
      operationId: getCacheEntries
      summary: Page through cache entries
      description: 404 {"detail":"Cache is disabled"} when the cache is off.
      parameters:
        - name: limit
          in: query
          schema: {type: integer, default: 200, minimum: 0, maximum: 1000}
        - name: offset
          in: query
          schema: {type: integer, default: 0, minimum: 0}
        - name: scope
          in: query
          schema: {type: string}
        - name: search
          in: query
          schema: {type: string}
        - name: expired_only
          in: query
          schema: {type: boolean, default: false}
      responses:
        '200':
          description: One page of entries.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/CacheEntriesPage'}
        '404':
          description: Cache is disabled.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Detail'}
        '422': {$ref: '#/components/responses/ValidationError'}
        '401': {$ref: '#/components/responses/Unauthorized'}
  /api/v1/debug/cache/entry/{key}:
    get:
      tags: [cache]
      operationId: getCacheEntry
      summary: One cache entry, decoded
      description: >-
        The key may contain slashes and colons and is matched greedily.
        404 when the cache is disabled or the key is unknown; 500 when the
        stored value cannot be read.
      parameters:
        - name: key
          in: path
          required: true
          schema: {type: string}
      responses:
        '200':
          description: Entry row plus decoded value and raw string.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/CacheEntryDetail'}
        '404':
          description: Cache disabled or key unknown.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Detail'}
        '500':
          description: Stored value unreadable.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Detail'}
        '401': {$ref: '#/components/responses/Unauthorized'}
  /api/v1/debug/cache/stats:
    get:
      tags: [cache]
      operationId: getCacheStats
      summary: Cache memory/DB gauges
      description: 404 {"detail":"Cache is disabled"} when the cache is off.
      responses:
        '200':
          description: Cache statistics.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/CacheStats'}
        '404':
          description: Cache is disabled.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Detail'}
        '401': {$ref: '#/components/responses/Unauthorized'}
  /api/v1/debug/consume-pause:
    get:
      tags: [debug-tools]
      operationId: getConsumePause
      summary: Consume-pause state
      description: >-
        Always 200 — `enabled: false` in the body is the signal the Live
        page hides its pause control on (no failure-probe needed).
      responses:
        '200':
          description: Current pause state and the configured presets.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ConsumePauseState'}
        '401': {$ref: '#/components/responses/Unauthorized'}
    post:
      tags: [debug-tools]
      operationId: postConsumePause
      summary: Pause consuming for a bounded period
      description: >-
        Pauses fetching on the assigned partitions for duration_seconds
        (1-3600), then auto-resumes. Never leaves the consumer group — the
        poll loop and heartbeats continue, so no rebalance is triggered. A
        pause while one is active replaces the deadline. 403 {"error":...}
        when ui.consume_pause.enabled=false (default — the feature is
        opt-in); 503 while the consumer is not running.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                duration_seconds: {type: integer, minimum: 1, maximum: 3600}
              required: [duration_seconds]
      responses:
        '200':
          description: The new (active) pause state.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ConsumePauseState'}
        '422': {$ref: '#/components/responses/ValidationError'}
        '503':
          description: Consumer not running yet.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Detail'}
        '401': {$ref: '#/components/responses/Unauthorized'}
  /api/v1/debug/consume-resume:
    post:
      tags: [debug-tools]
      operationId: postConsumeResume
      summary: Resume consuming now
      description: >-
        Ends an active debug pause immediately. Idempotent — resuming while
        not paused answers 200 with the inactive state. Partitions still
        held by backpressure stay paused until queues drain (the pause
        never overrides backpressure). 403 when
        ui.consume_pause.enabled=false.
      responses:
        '200':
          description: The (inactive) pause state.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ConsumePauseState'}
        '401': {$ref: '#/components/responses/Unauthorized'}
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: >-
        Only enforced when ui.auth_token is configured; otherwise every
        route is open (private-contour default). The token may also ride as
        a ?token= query parameter (downloads, WebSocket).
  responses:
    Unauthorized:
      description: Missing or invalid bearer token (only when auth is configured).
      content:
        application/json:
          schema: {$ref: '#/components/schemas/Detail'}
    ValidationError:
      description: Invalid parameters or body.
      content:
        application/json:
          schema: {$ref: '#/components/schemas/ValidationDetail'}
  schemas:
...
    Detail:
      description: Standard error envelope for 401/404/500/503.
      type: object
      properties:
        detail: {type: string}
      required: [detail]
    ValidationDetail:
      description: Pinned-minimal 422 envelope — consumers read only msg.
      type: object
      properties:
        detail:
          type: array
          items:
            type: object
            properties:
              loc:
                type: array
                items: {}
              msg: {type: string}
            required: [loc, msg]
      required: [detail]
...
    CacheEntry:
      type: object
      properties:
        key: {type: string}
        scope: {type: [string, 'null']}
        value:
          type: string
          description: Raw stored string.
        size_bytes: {type: integer}
        created_at_ms: {type: integer}
        updated_at_ms: {type: integer}
        expires_at_ms: {type: [integer, 'null']}
        origin_worker_id: {type: [string, 'null']}
      required: [key, size_bytes, created_at_ms, updated_at_ms]
    CacheEntriesPage:
      type: object
      properties:
        entries:
          type: array
          items: {$ref: '#/components/schemas/CacheEntry'}
        total: {type: integer}
        limit: {type: integer}
        offset: {type: integer}
      required: [entries, total, limit, offset]
    CacheEntryDetail:
      allOf:
        - {$ref: '#/components/schemas/CacheEntry'}
        - type: object
          properties:
            value:
              description: Decoded JSON value (null when not decodable).
            raw_value: {type: string}
          required: [raw_value]
    CacheStats:
      type: object
      properties:
        entries_in_memory: {type: integer}
        bytes_in_memory: {type: integer}
        entries_in_db: {type: integer}
        bytes_in_db: {type: integer}
      required: [entries_in_memory, bytes_in_memory, entries_in_db, bytes_in_db]
    ConsumePauseState:
      type: object
      description: >-
        State of the operator debug pause. resume_at_ms / requested_seconds
        are null while inactive; durations_seconds carries the configured
        Live-page presets.
      properties:
        enabled: {type: boolean}
        durations_seconds:
          type: array
          items: {type: integer}
        active: {type: boolean}
        resume_at_ms: {type: [integer, 'null']}
        requested_seconds: {type: [integer, 'null']}
      required: [enabled, durations_seconds, active, resume_at_ms, requested_seconds]
...